+++ lib/lsp/lspserver.ts
@@ -6,10 +6,13 @@
 
 import {
     createConnection, TextDocuments, TextDocument, Diagnostic, DiagnosticSeverity,
-    ProposedFeatures, InitializeParams, Proposed, Range, DocumentFormattingParams, CompletionItem,
-    TextDocumentPositionParams, Definition, Hover, Location, MarkedString
+    ProposedFeatures, InitializeParams, InitializeResult, DidChangeConfigurationNotification, Range, DocumentFormattingParams, CompletionItem,
+    TextDocumentPositionParams, Definition, Hover, Location, MarkedString,
+	ReferenceParams, CodeLensParams, CodeLens, Command, ExecuteCommandParams, Position
 } from 'vscode-languageserver';
 
+import URI from 'vscode-uri';
+ 
 import { ErlangLspConnection, ParsingResult } from './erlangLspConnection';
 import { ErlangShellLSP } from './ErlangShellLSP';
 import { IErlangShellOutput } from '../GenericShell';
@@ -46,7 +49,7 @@ let module2helpPage: Map<string, string[]> = new Map();
 //trace for debugging 
 let traceEnabled = false;
 
-connection.onInitialize(async (params: InitializeParams) => {
+connection.onInitialize(async (params: InitializeParams): Promise<InitializeResult> => {
 
     //connection.console.log("onInitialize.");
     
@@ -61,18 +64,24 @@ connection.onInitialize(async (params: InitializeParams) => {
 
     // Does the client support the `workspace/configuration` request? 
     // If not, we will fall back using global settings
-    hasWorkspaceFolderCapability = (capabilities as Proposed.WorkspaceFoldersClientCapabilities).workspace && !!(capabilities as Proposed.WorkspaceFoldersClientCapabilities).workspace.workspaceFolders;
-    hasConfigurationCapability = (capabilities as Proposed.ConfigurationClientCapabilities).workspace && !!(capabilities as Proposed.ConfigurationClientCapabilities).workspace.configuration;
+    hasWorkspaceFolderCapability = capabilities.workspace && !!capabilities.workspace.workspaceFolders;
+    hasConfigurationCapability = capabilities.workspace && !!capabilities.workspace.configuration;
     if (hasConfigurationCapability) {
-        debugLog(JSON.stringify((capabilities as Proposed.ConfigurationClientCapabilities).workspace.configuration));
+        debugLog(JSON.stringify(capabilities.workspace.configuration));
     }
     debugLog(`capabilities => hasWorkspaceFolderCapability:${hasWorkspaceFolderCapability}, hasConfigurationCapability:${hasConfigurationCapability}`);
-    return {
+	//return new InitializeResult()
+    return <InitializeResult>{
         capabilities: {
             textDocumentSync: documents.syncKind,
             documentFormattingProvider : true,
             definitionProvider: true,
-            hoverProvider: true
+            hoverProvider: true,
+			codeLensProvider :  { resolveProvider : true },
+			referencesProvider : true,
+			// executeCommandProvider: {
+			// 	commands : ["erlang.showReferences"]
+			// },
             // completionProvider : {
             //  resolveProvider: true,
             //  triggerCharacters: [ ':' ]
@@ -85,10 +94,13 @@ connection.onInitialized(async () => {
     var globalConfig = await connection.workspace.getConfiguration("erlang");
     if (globalConfig) {
         let erlangConfig = globalConfig;
-        if (erlangConfig && erlangConfig.languageServerProtocol.verbose) {
+        if (erlangConfig && erlangConfig.verbose) {
             traceEnabled = true;
         }
     }
+	if (hasConfigurationCapability) {
+		connection.client.register(DidChangeConfigurationNotification.type, undefined);
+	}
 
     //debugLog("connection.onInitialized");
     if (hasWorkspaceFolderCapability) {
@@ -98,20 +110,26 @@ connection.onInitialized(async () => {
     }
 });
 
+connection.onExecuteCommand((cmdParams: ExecuteCommandParams): any => {
+	debugLog(`onExecuteCommand : ${JSON.stringify(cmdParams)}`);
+	//connection.sendRequest(CommandReques)
+	return null;
+});
+
 connection.onShutdown(() => {
     debugLog("connection.onShutDown");
-    erlangLsp.Kill();
+	erlangLspConnection.Quit();
 });
 
 connection.onExit(() => {
     debugLog("connection.onExit");
-    erlangLsp.Kill();
+	erlangLspConnection.Quit();
 });
 
 // The global settings, used when the `workspace/configuration` request is not supported by the client.
 // Please note that this is not the case when using this server with the client provided in this example
 // but could happen with other clients.
-const defaultSettings: ErlangSettings = { erlangPath: "", rebarBuildArgs:[],  rebarPath: "",  languageServerProtocol : { verbose : false} };
+const defaultSettings: ErlangSettings = { erlangPath: "", rebarBuildArgs:[],  rebarPath: "", linting: true, codeLensEnabled: false, verbose: false };
 let globalSettings: ErlangSettings = defaultSettings;
 
 // Cache the settings of all open documents
@@ -126,7 +144,11 @@ connection.onDidChangeConfiguration(change => {
     }
 
     // Revalidate all open text documents
-    documents.all().forEach(validateTextDocument);
+    documents.all().forEach(document => {
+		let diagnostics: Diagnostic[] = [];
+		connection.sendDiagnostics({ uri: document.uri, diagnostics });
+		validateTextDocument(document);
+	});
 });
 
 function getDocumentSettings(resource: string): Thenable<ErlangSettings> {
@@ -141,8 +163,22 @@ function getDocumentSettings(resource: string): Thenable<ErlangSettings> {
     return result;
 }
 
-// Only keep settings for open documents
+documents.onDidOpen(e => {
+	var validateWhenReady = function () {
+		if (erlangLspConnection.isConnected)
+		    validateTextDocument(e.document);
+		else {
+			setTimeout(function () {
+				validateWhenReady();
+			}, 100);
+		}
+	};
+	validateWhenReady();
+});
+	
 documents.onDidClose(e => {
+	let diagnostics: Diagnostic[] = [];
+	connection.sendDiagnostics({ uri: e.document.uri, diagnostics });
     documentSettings.delete(e.document.uri);
     erlangLspConnection.onDocumentClosed(e.document.uri);
 });
@@ -154,6 +190,10 @@ documents.onDidChangeContent((change) => {
 });
 
 async function validateTextDocument(textDocument: TextDocument): Promise<void> {
+	var erlangConfig = await connection.workspace.getConfiguration("erlang");
+	if (erlangConfig && !erlangConfig.linting) {
+		return;
+	}
     // In this simple example we get the settings for every validate run.
     let settings = await getDocumentSettings(textDocument.uri);
     // .hrl files show incorrect errors e.g. about not used records, ddisabled for now
@@ -182,16 +222,22 @@ connection.onDefinition(async (textDocumentPosition: TextDocumentPositionParams)
 
 function markdown(str: string): string {
     str = str.trim();
-    var reg = /(((< *\/[^>]+>)|(<[^>]+>))|([^<]+))/g;
+    var reg = /(((< *\/[^>]+>)|(< *([a-zA-Z0-9]+)[^>]*>))|([^<]+))/g;
     var out = '';
     var result;
     var tags = [];
     var off = [];
     while ((result = reg.exec(str)) !== null) {
         if (result[4]) {
+            var tagName = result[5];            
             var tag = ''
             var endTag = '';
-            if (result[4].indexOf('<p') >= 0 || result[4].indexOf('name="') >= 0) {
+            if (tagName === 'br') {
+                if (off.length === 0)
+                    out += '  \n';
+                continue;
+            }
+            else if (tagName === 'p' || result[4].indexOf('name="') >= 0) {
                 tag = '';
                 endTag = '  \n';
             }
@@ -220,65 +266,124 @@ function markdown(str: string): string {
             else
                 out += top;
         }
-        else if (result[5] && off.length === 0)
-            out += result[5];
+        else if (result[6] && off.length === 0)
+            out += result[6];
     }
     return out;
 }
 
 async function getModuleHelpPage(moduleName: string): Promise<string[]> {
-	if (module2helpPage.has(moduleName)) {
-		return module2helpPage.get(moduleName);
-	}
-	else {
-		return new Promise<string[]>(resolve => {
-			http.get('http://erlang.org/doc/man/' + moduleName + '.html', (response) => {
-				let contents:string = '';
-				response.on('data', (chunk) => {
-					contents += chunk;
-				});
-				response.on('end', () => {
-					module2helpPage.set(moduleName, contents.split('\n'));
-					resolve(module2helpPage.get(moduleName));
-				});   
-			}).on("error", (error) => {
-				module2helpPage.set(moduleName, []);
-				resolve([]);
-			});		  
-		});
-	}
+    if (module2helpPage.has(moduleName)) {
+        return module2helpPage.get(moduleName);
+    }
+    else {
+        return new Promise<string[]>(resolve => {
+            http.get('http://erlang.org/doc/man/' + moduleName + '.html', (response) => {
+                let contents:string = '';
+                response.on('data', (chunk) => {
+                    contents += chunk;
+                });
+                response.on('end', () => {
+                    module2helpPage.set(moduleName, contents.split('\n'));
+                    resolve(module2helpPage.get(moduleName));
+                });   
+            }).on("error", (error) => {
+                module2helpPage.set(moduleName, []);
+                resolve([]);
+            });       
+        });
+    }
 };
 
 function extractHelpForFunction(functionName: string, htmlLines: string[]): string {
-	var helpText: string = '';
-	var found = false;
-	for (var i = 0; i < htmlLines.length; ++i) {
-		var trimmed = htmlLines[i].trim();
-		if (!found) {
-			if (trimmed.indexOf('name="' + functionName) >= 0) {
-				found = true;
-				helpText = trimmed;
-			}
-		}
-		else {
-			if (!trimmed || trimmed.indexOf('name="') !== trimmed.indexOf('name="' + functionName))
-				break;
-			else
-				helpText += '\n' + trimmed;
-		}
-	}
-	return helpText;
+    var helpText: string = '';
+    var found = false;
+    for (var i = 0; i < htmlLines.length; ++i) {
+        var trimmed = htmlLines[i].trim();
+        if (!found) {
+            if (trimmed.indexOf('name="' + functionName) >= 0) {
+                found = true;
+                helpText = trimmed;
+            }
+        }
+        else {
+            if (!trimmed || trimmed.indexOf('name="') !== trimmed.indexOf('name="' + functionName))
+                break;
+            else
+                helpText += '\n' + trimmed;
+        }
+    }
+    return helpText;
 }
 
 connection.onHover(async (textDocumentPosition: TextDocumentPositionParams): Promise<Hover> => {
-	var uri = textDocumentPosition.textDocument.uri;
+    var uri = textDocumentPosition.textDocument.uri;
     let res = await erlangLspConnection.getHoverInfo(uri, textDocumentPosition.position.line, textDocumentPosition.position.character);
     if (res) {
-		var htmlLines = await getModuleHelpPage(res.moduleName);
-		return {contents: markdown(extractHelpForFunction(res.functionName, htmlLines))};
+		debugLog(JSON.stringify(res))
+		if (res.text) {
+	        return {contents: res.text};
+		}
+		else {
+			var htmlLines = await getModuleHelpPage(res.moduleName);
+        	return {contents: markdown(extractHelpForFunction(res.functionName, htmlLines))};
+		}
     }
     return null;
 });
+
+connection.onReferences(async (reference : ReferenceParams) : Promise<Location[]> => {
+    var uri = reference.textDocument.uri;
+    let res = await erlangLspConnection.getReferencesInfo(uri, reference.position.line, reference.position.character);
+	if (res) {
+		var Result = new Array<Location>();
+		res.forEach(ref => {
+			Result.push(Location.create(ref.uri, Range.create(ref.line, ref.character, ref.line, ref.character)));
+		});
+		return Result;
+	}
+	return null;
+});
+
+connection.onCodeLens(async (codeLens: CodeLensParams) : Promise<CodeLens[]>  => {
+	var erlangConfig = await connection.workspace.getConfiguration("erlang");
+	if (erlangConfig) {
+		if (!erlangConfig.codeLensEnabled) {
+			return [];
+		}
+	}
+    var uri = codeLens.textDocument.uri;
+    let res = await erlangLspConnection.getCodeLensInfo(uri);
+	if (res) {		
+		var Result = new Array<CodeLens>();
+		res.codelens.forEach(ref => {
+			if (ref.data.exported) {
+				let exportedCodeLens = CodeLens.create(Range.create(ref.line, ref.character, ref.line, ref.character + ref.data.func_name.length), ref.data);
+				exportedCodeLens.command = Command.create("exported", "");
+				Result.push(exportedCodeLens);
+			}
+			if (!ref.data.exported || ref.data.count > 0) {
+				let codeLens = CodeLens.create(Range.create(ref.line, ref.character, ref.line, ref.character + ref.data.func_name.length), ref.data);
+				//codeLens.command = null; //set to null to invoke OnCodeLensResolve
+				codeLens.command = ref.data.count == 0 ? Command.create("unused","") : 
+					Command.create(`${ref.data.count} private references`, "editor.action.findReferences", 
+										URI.parse(res.uri), {lineNumber : ref.line+1, column:ref.character+1});
+				Result.push(codeLens);	
+			}
+		});
+		return Result;
+	}
+	return null;
+});
+
+connection.onCodeLensResolve(async (codeLens : CodeLens) : Promise<CodeLens> => {	
+
+	// let command = Command.create(`${codeLens.data.count} private references`, "erlang.showReferences", 
+	// 						 codeLens.data.uri, codeLens.range.start, []);
+	codeLens.command = Command.create("onCodeLensResolve","");
+	return codeLens;
+});
+
 
 //https://stackoverflow.com/questions/38378410/can-i-add-a-completions-intellisense-file-to-a-language-support-extension
 connection.onCompletion(async (textDocumentPosition: TextDocumentPositionParams):Promise<CompletionItem[]> => {